Create a Plasma sub-package.
Apparently Plasma uses /usr/share/wallpapers, so create a sub-package with /usr/share/wallpapers/solus, which is a symlink to /usr/share/backgrounds/solus.

Having it as a sub-package also means we can put it in desktop.kde and therefore is more likely to be installed if a user is following our Desktop Environments documentation to install Plasma.